Move duplicated UI components to shared-composeui
